src/domain/services/test_OLD_DO_NOT_USE/base_test.py
"""Base test service implementation.

This module implements base testing functionality.
It provides:
- Test fixtures
- Common assertions
- Test utilities
- Mock factories
"""
from datetime import datetime, timedelta
from typing import Any, Dict, List, Optional, Type, Union
import json
import pytest
import uuid

from src.domain.services.common.base import BaseService

class BaseTestService(BaseService):
    """Base service for testing functionality.
    
    This service is responsible for:
    - Managing test fixtures
    - Providing assertions
    - Test utilities
    - Mock object creation
    """
    
    def __init__(
        self,
        mock_factory: Optional['MockFactory'] = None,
        fixture_path: Optional[str] = None
    ):
        """Initialize base test service.
        
        Args:
            mock_factory: Optional mock factory
            fixture_path: Optional fixture directory path
        """
        super().__init__()
        self._mock_factory = mock_factory
        self._fixture_path = fixture_path
        self._fixtures: Dict[str, Any] = {}
        self._mocks: Dict[str, Any] = {}
    
    def create_mock(
        self,
        class_type: Type,
        **kwargs
    ) -> Any:
        """Create mock object.
        
        Args:
            class_type: Class to mock
            **kwargs: Mock attributes
            
        Returns:
            Mock object
        """
        self._log_entry(
            "create_mock",
            class_type=class_type.__name__
        )
        
        try:
            # Use mock factory if available
            if self._mock_factory:
                mock = self._mock_factory.create(
                    class_type,
                    **kwargs
                )
            else:
                # Create basic mock
                mock = type(
                    f"Mock{class_type.__name__}",
                    (object,),
                    kwargs
                )()
            
            # Store mock
            mock_id = str(uuid.uuid4())
            self._mocks[mock_id] = mock
            
            return mock
            
        except Exception as e:
            self._log_error("create_mock", e)
            raise
    
    def load_fixture(
        self,
        name: str,
        **kwargs
    ) -> Any:
        """Load test fixture.
        
        Args:
            name: Fixture name
            **kwargs: Fixture parameters
            
        Returns:
            Fixture data
            
        Raises:
            ValueError: If fixture not found
        """
        self._log_entry("load_fixture", name=name)
        
        try:
            # Check cache
            if name in self._fixtures:
                return self._fixtures[name]
            
            if not self._fixture_path:
                raise ValueError("Fixture path not set")
            
            # Load fixture file
            fixture_file = f"{self._fixture_path}/{name}.json"
            with open(fixture_file, "r") as f:
                data = json.load(f)
            
            # Apply parameters
            if kwargs:
                data = self._apply_params(data, kwargs)
            
            # Cache fixture
            self._fixtures[name] = data
            
            return data
            
        except Exception as e:
            self._log_error("load_fixture", e)
            raise ValueError(f"Failed to load fixture: {str(e)}")
    
    def assert_equal(
        self,
        actual: Any,
        expected: Any,
        msg: Optional[str] = None
    ) -> None:
        """Assert values are equal.
        
        Args:
            actual: Actual value
            expected: Expected value
            msg: Optional message
            
        Raises:
            AssertionError: If values not equal
        """
        self._log_entry(
            "assert_equal",
            actual=actual,
            expected=expected
        )
        
        try:
            assert actual == expected, (
                msg or f"Expected {expected}, got {actual}"
            )
            
        except AssertionError as e:
            self._log_error("assert_equal", e)
            raise
    
    def assert_not_equal(
        self,
        actual: Any,
        expected: Any,
        msg: Optional[str] = None
    ) -> None:
        """Assert values are not equal.
        
        Args:
            actual: Actual value
            expected: Expected value
            msg: Optional message
            
        Raises:
            AssertionError: If values equal
        """
        self._log_entry(
            "assert_not_equal",
            actual=actual,
            expected=expected
        )
        
        try:
            assert actual != expected, (
                msg or f"Expected not {expected}"
            )
            
        except AssertionError as e:
            self._log_error("assert_not_equal", e)
            raise
    
    def assert_true(
        self,
        value: Any,
        msg: Optional[str] = None
    ) -> None:
        """Assert value is true.
        
        Args:
            value: Value to check
            msg: Optional message
            
        Raises:
            AssertionError: If value not true
        """
        self._log_entry("assert_true", value=value)
        
        try:
            assert value, msg or "Expected True"
            
        except AssertionError as e:
            self._log_error("assert_true", e)
            raise
    
    def assert_false(
        self,
        value: Any,
        msg: Optional[str] = None
    ) -> None:
        """Assert value is false.
        
        Args:
            value: Value to check
            msg: Optional message
            
        Raises:
            AssertionError: If value not false
        """
        self._log_entry("assert_false", value=value)
        
        try:
            assert not value, msg or "Expected False"
            
        except AssertionError as e:
            self._log_error("assert_false", e)
            raise
    
    def assert_raises(
        self,
        exc_type: Type[Exception],
        func: callable,
        *args,
        **kwargs
    ) -> None:
        """Assert function raises exception.
        
        Args:
            exc_type: Expected exception type
            func: Function to call
            *args: Function arguments
            **kwargs: Function keyword arguments
            
        Raises:
            AssertionError: If exception not raised
        """
        self._log_entry(
            "assert_raises",
            exc_type=exc_type.__name__
        )
        
        try:
            with pytest.raises(exc_type):
                func(*args, **kwargs)
                
        except Exception as e:
            self._log_error("assert_raises", e)
            raise
    
    def assert_dict_equal(
        self,
        actual: Dict,
        expected: Dict,
        msg: Optional[str] = None
    ) -> None:
        """Assert dictionaries are equal.
        
        Args:
            actual: Actual dictionary
            expected: Expected dictionary
            msg: Optional message
            
        Raises:
            AssertionError: If dictionaries not equal
        """
        self._log_entry(
            "assert_dict_equal",
            actual=actual,
            expected=expected
        )
        
        try:
            # Check keys match
            self.assert_equal(
                set(actual.keys()),
                set(expected.keys()),
                "Dictionary keys don't match"
            )
            
            # Check values match
            for key in expected:
                self.assert_equal(
                    actual[key],
                    expected[key],
                    f"Value mismatch for key '{key}'"
                )
                
        except AssertionError as e:
            self._log_error("assert_dict_equal", e)
            raise
    
    def assert_list_equal(
        self,
        actual: List,
        expected: List,
        msg: Optional[str] = None
    ) -> None:
        """Assert lists are equal.
        
        Args:
            actual: Actual list
            expected: Expected list
            msg: Optional message
            
        Raises:
            AssertionError: If lists not equal
        """
        self._log_entry(
            "assert_list_equal",
            actual=actual,
            expected=expected
        )
        
        try:
            # Check lengths match
            self.assert_equal(
                len(actual),
                len(expected),
                "List lengths don't match"
            )
            
            # Check items match
            for i, (a, e) in enumerate(zip(actual, expected)):
                self.assert_equal(
                    a,
                    e,
                    f"Item mismatch at index {i}"
                )
                
        except AssertionError as e:
            self._log_error("assert_list_equal", e)
            raise
    
    def assert_almost_equal(
        self,
        actual: float,
        expected: float,
        places: int = 7,
        msg: Optional[str] = None
    ) -> None:
        """Assert floats are almost equal.
        
        Args:
            actual: Actual float
            expected: Expected float
            places: Decimal places to check
            msg: Optional message
            
        Raises:
            AssertionError: If floats not almost equal
        """
        self._log_entry(
            "assert_almost_equal",
            actual=actual,
            expected=expected,
            places=places
        )
        
        try:
            assert round(abs(actual - expected), places) == 0, (
                msg or f"Expected {expected}, got {actual}"
            )
            
        except AssertionError as e:
            self._log_error("assert_almost_equal", e)
            raise
    
    def _apply_params(
        self,
        data: Any,
        params: Dict
    ) -> Any:
        """Apply parameters to fixture data.
        
        Args:
            data: Fixture data
            params: Parameters to apply
            
        Returns:
            Updated data
        """
        if isinstance(data, dict):
            return {
                k: self._apply_params(v, params)
                for k, v in data.items()
            }
        
        if isinstance(data, list):
            return [
                self._apply_params(v, params)
                for v in data
            ]
        
        if isinstance(data, str):
            # Replace parameters in strings
            for k, v in params.items():
                data = data.replace(f"${k}", str(v))
            
        return data
